Added parameter "use.prepared.tables" to config.

+++ src/src/main/java/com/tsurugidb/benchmark/phonebill/db/iceaxe/dao/DdlIceaxe.java
@@ -19,9 +19,11 @@ import com.tsurugidb.iceaxe.transaction.exception.TsurugiTransactionRuntimeExcep
  */
 public class DdlIceaxe implements Ddl {
 	private final PhoneBillDbManagerIceaxe manager;
+	private final boolean usePreparedTables;
 
 	public DdlIceaxe(PhoneBillDbManagerIceaxe phoneBillDbManagerIceaxe) {
 		this.manager = phoneBillDbManagerIceaxe;
+		this.usePreparedTables = manager.getConfig().usePreparedTables;
 	}
 
 	@Override
@@ -33,7 +35,11 @@ public class DdlIceaxe implements Ddl {
 			throw new UncheckedIOException(e);
 		}
 		if (opt.isPresent()) {
-			execute("drop table " + tableName);
+			if (usePreparedTables && tableName.equalsIgnoreCase("history")) {
+				truncateTable(tableName);
+			} else {
+				execute("drop table " + tableName);
+			}
 		}
 	}
 
@@ -44,17 +50,21 @@ public class DdlIceaxe implements Ddl {
 
 	@Override
 	public void createHistoryTable() {
-		String create_table = "create table history ("
-				+ "caller_phone_number varchar(15) not null," 		// 発信者電話番号
-				+ "recipient_phone_number varchar(15) not null," 	// 受信者電話番号
-				+ "payment_categorty char(1) not null," 			// 料金区分
-				+ "start_time timestamp not null,"			 		// 通話開始時刻
-				+ "time_secs int not null," 					// 通話時間(秒)
-				+ "charge int," 								// 料金
-				+ "df int not null," 							// 論理削除フラグ
-				+ "primary key (caller_phone_number, start_time)"
-				+ ")";
-		execute(create_table);
+		if (usePreparedTables) {
+			truncateTable("history");
+		} else {
+			String create_table = "create table history ("
+					+ "caller_phone_number varchar(15) not null," 		// 発信者電話番号
+					+ "recipient_phone_number varchar(15) not null," 	// 受信者電話番号
+					+ "payment_categorty char(1) not null," 			// 料金区分
+					+ "start_time timestamp not null,"			 		// 通話開始時刻
+					+ "time_secs int not null," 					// 通話時間(秒)
+					+ "charge int," 								// 料金
+					+ "df int not null," 							// 論理削除フラグ
+					+ "primary key (caller_phone_number, start_time)"
+					+ ")";
+			execute(create_table);
+		}
 	}
 
 	public void createContractsTable() {
